requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
defaults==0.3.0
et-xmlfile==1.1.0
execnet==1.9.0
Faker==8.8.2
gevent==21.1.2
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
jsonpath==0.82
loguru==0.5.3
openpyxl==3.0.7
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-rerunfailures==10.0
pytest-xdist==2.3.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typeguard==2.13.3
urllib3==1.26.6
zope.event==4.5.0
zope.interface==5.4.0
